+++ tests/integration/components/contributor-list/component-test.ts
@@ -6,13 +6,22 @@ import hbs from 'htmlbars-inline-precompile';
 
 const i18nStub = Service.extend({
     translations: EmberObject.create({
-        general: EmberObject.create({
-            and: 'and',
-            more: 'more',
-        }),
+        list: {
+            two_item: {
+                delimiter: ' and ',
+            },
+            many_item: {
+                first_delimiter: ', ',
+                each_delimiter: ', ',
+                last_delimiter: ', and ',
+            },
+        },
     }),
 
-    t(key: string): string {
+    t(key: string, options: any): string {
+        if (key === 'contributor_list.and_x_more') {
+            return `${options.get('x')} more`;
+        }
         // @ts-ignore
         return this.get('translations').get(key);
     },