+++ tests/unit/guid-node/forks/controller-test.ts
@@ -1,0 +1,12 @@
+import { setupTest } from 'ember-qunit';
+import { module, test } from 'qunit';
+
+module('Unit | Controller | guid-node/forks', hooks => {
+    setupTest(hooks);
+
+    // Replace this with your real tests.
+    test('it exists', function(assert) {
+        const controller = this.owner.lookup('controller:guid-node/forks');
+        assert.ok(controller);
+    });
+});