+++ types/ember-toastr/index.d.ts
@@ -2,7 +2,8 @@ declare module 'ember-toastr/services/toast' {
     import Service from '@ember/service';
 
     export default class Toast extends Service {
-        success(message: string): void;
-        error(message: string): void;
+        success(message: string, title?: string): void;
+        error(message: string, title?: string): void;
+        info(message: string, title?: string): void;
     }
 }